(PE-15435) Replace wherefore with cmd
The support script had a handful of locations where the wherefore function from
the utilities library was used to test for command existance. This patch
replaces the wherefore function with a standard cmd function which is a simple
wrapper around the shell builtin hash that can be used to test for existance
without the need to suppress output.