(PE-15435) Remove installer exception handler
The exception handler executes some rollback and cleanup logic for the
installer, finalizes some logfiles, and then displays an error message. The
support script doesn't use any of the state that is cleaned up or rolled back
and doesn't write to the installer logfile, so the only benefit we get out of
this is an error message which shows up if the script is canceled with CTRL-C.

This patch removes the exception handlers as the error message isn't worth the
additional complexity.