(PE-15945) Only gather whitelisted config files
Prior to this patch, the gather_enterprise_files function used a tar pipe to
copy /etc/puppetlabs into the support script output directory with a blacklist
of exclusions.  The problem with this approach is that sensitive material, such
as SSL keys, can easily escape the blacklist due to lack of maintenance or
users creating backup copies under /etc/puppetlabs.

This patch re-factors the gather_enterprise_files function to run a plain copy
on a whitelist of configuration files.