(PE-16537) Collect Puppet Server environments API output
This patch adds a puppetserver_environments check that collects the output of
the `puppet/v3/environments` endpoint of the Puppet Server. This endpoint
reports the list of environments visible to the server along with the effective
values of environment.conf settings. This information is useful for debugging
code loading and classifier sync issues.