(PE-17822) Move puppet enterprise face from pe_manager
Prior to this commit the `pe_support_script` module had a depedency on
`puppetlabs-pe_manager` which implemented the `puppet enterprise` face,
this is being renamed to `puppet infrastructure` with `puppet enterprise
support` remaining unchanged.

This commit removes the dependecy on `pe_manager` and implements the
`enterprise` face entirely within this module.