(PE-16091) Disable support script on non-master platforms
This patch disables the support script on platforms that are not running a
variant of RedHat, Debian or SUSE operating systems. This primarily guards
against undefined and untested behavior occuring on Solaris, AIX and OS X.
These platforms may be enabled on a future date when test coverage
catches up.