(maint) Update archive creation
This patch updates the creation of support script archives to combine
`tar` and `gzip` using a pipeline. This construction prevents redundant
copies of the support data from being created and using up disk space. A
umask is also applied to ensure that the resulting tarball is only
readable by the root user.