(PE-19402) Update test dependencies
The test dependencies inherited from the module skeleton included support for
fairly old versions of puppet, which means that some of the gem versions used
were starting to break due to projects dropping support for Ruby 1.9. This
patch updates the Gemfile to use Puppet 4.5/PE 2016.2.0 as the new minimum
version.