(PE-15953) Gather debug output from facter
This commit updates facter data output to include debug information.
The main facter data is split into a separate json file and the
debug data is kept in a separate file.  The code is a little messy
to ensure only common UNIX tools are used and no additional ruby
overhead is incurred.