(PE-7029) Add classifier information to support script
Prior to this commit, the support script did not gather any classification
data.  This complicates debugging customer issues that appear when
reviewing the rest of the data contained in the output.  This commit
adds a --classifier flag to the puppet enterprise support face that
gathers the classifier's group endpoint if it is run on a system
with pe-console-services installed.

Closes puppetlabs/puppetlabs-pe_support_script#27

Signed-off-by: Charlie Sharpsteen <chuck@puppet.com>