(PE-9661) Capture structured module information
Prior to this commit, we only gathered the raw CLI output of
puppet module list, which is difficult to programmatically parse.
With this commit, we now dump a YAML version of the data as well.
JSON would have been better, but --render-as json does not include
the module's version number in the output.  The YAML output was far
more descriptive.