(PE-20635) Remove dependency list from metadata.json
Since PE modules are now delivered as a package (pe-modules) in all PE
lines, and since you can't separately install these packages by the
forge, the dependency metadata is not providing any benefit and is
causing a lot of upkeep churn around releases.

So we're removing it...

In Puppet 5 the SemanticPuppet gem changed how it handles semantic
version checks, and warnings began to surface testing `puppet module
list --strict-semver` when testing development builds of PE.  The issue
was with dependency checks between PE modules.  Our packaging systems
overwrite the "version" field of metadata.json with the git describe of
the sha being pulled into a particular pe-modules package version
shipped in PE.  Due to the fact that we keep code synchronized across
multiple branches of modules by merging up across branches, it is often
necessary to include rc tags (such as 2017.3.0-rc0) on a module branch
after having merged up from an earlier branch with a more recent tag, so
as to ensure that we have a git describe that our packaging system
considers newer (otherwise it will prevent promotion).

The new SemanticPuppet does not consider an rc version (such as
2017.3.0-rc0-5-gabcde) to be part of a semver sequence.  A test such as
>= 2017.2.0 will not conclude that 2017.3.0-rc0-* is greater.  You +can+
ensure that rc versions are evaluated by specifing >= 2017.3.0-0, but it
will only consider a patch if major, minor, patch matches.  So a version
such as 2017.4.0-rc1-* will not be considered >=2017.3.0-0.

To keep the PE module metadata dependency lists in a state where they
would not show any 'invalid' warnings during a `puppet module list
--strict-semver` would require maintaining separate branches for each
release line with the correct x.y.z-r, which would generate branches on
many modules that were used only to track this metadata change.