(PE-16002) Simplify db_size_from_du in support script
Prior to this commit, if the postgresql database was inaccessible,
the support script would fail to pull the database sizes from du.
This was due to the fact that we queried the database to pull the
oids of the databases to determine their directory names.

Since it doesn't make a lot of sense to not pull filesystem information
simply because the DB is dead, we should skip the extra database
hit and just run du against the established database directory
locations.  They are unlikely to change, and this way we get the data
no matter what.