(PE-13070) Collect modules for each environment in 2016.4
Prior to this commit, we collected the module information from
only the production environment.  When debugging issues, it is
not uncommon for us to need to know the modules installed in other
environments.

With this commit, we continue to gather the data as before, but
a new "modules" directory has been added to hold per-environment
module listings.

In 2017.3, this will be replaced with a simple call to the
environment_modules endpoint.