(PE-13070) Switch to environment_modules endpoint for module info (#39)
Prior to this commit, we used a puppet module list command
to pull module data.  This would only capture modules in the
production environment and produced output that was difficult to parse
programtically.  We later added yaml output as part of PE-9661,
but the output was still not ideal and continued to lack other
environments.

With this commit, we now use the Puppet 5 environment_modules endpoint,
which pulls modules from all environments and outputs in json.
Thus solving both issues.