(PE-24003) Streamline calls to puppet config print to reduce overhead
Prior to this commit, we had several duplicate calls to puppet config
print to pull cert related settings.  This meant additional invocations
of puppet, which led to unnecessary delays waiting for ruby to spin
up subshells.

With this commit, there is now a common function to use to pull
data from puppet config that caches the responses.  This ensures
that additional duplicate calls now and in the future will only
trigger a puppet invocation when required.