(PE-24023) Add option to limit log file collection
This commit adds a `--log-age` option to `puppet enterprise support` and an
associated `-l` flag to the bash script. These options accept a number, which
is interpreted as the maximum age in days for collecting logfiles. The string
"all" may also be passed to disable this collection limit. A default max age of
14 days is used.

The limit is enforced by using by passing a `-mtime` predicate to `find`
and by passing the `--since` flag to `journalctl`.