(PE-24078) Tag all LSB platforms as SLES
This commit adjusts the logic in the detect_platform function to recognize any
"Distributor ID" string returned by `lsb_release` that starts with "suse" as
being an indicator of the SLES platform. Previously, this function only
recognized the string "suse linux" which caused the support script to bail
on SLES 12.2 systems with the lsb-release package installed as the "linux"
suffix was dropped in 12.2.