(maint) Change rubocop profile to cleanups_only
Prior to this commit, the enabled profile was set to strict, which
enforces rubocop policies that we do not want in this repo. This commit
changes the rubocop.yml in the .sync.yml to use the cleanups_only
profile. The result of this change is that multiple checks will be
disabled and the rubocop check is reenabled in the .travis.yml