(SLV-653) Add a script for generating system stats
The script is intended to be used as part of an optionial feature to gate system stats the same way other stats
are gathered and managed.
It uses sar which will poll every x seconds and generate an average, which can then be captured into a file at the
same interval as the puppet-metrics-collector polling interval.  This keeps the total data down while not missing
things that cause short duration spikes (like compiles that last 6 seconds).  Initial use will be in performace testing.

Had to disable GetText/DecorateString cop for the entire script in the rubocop.yml because inline disables aren't working for some reason.  And rubocop is parsing for ruby 2.1 where the fix works, but the same fix doesn't work with later versions of ruby, like the one installed by puppet...