(SLV-771) Fix bug where process data is missed
The process data gathering greps process names for the ones we wanted to track... then passes the list to sar to track. If a process passed to sar exists when sar starts, but exits before sar gathers the first round of data, then sar will crash while printing the average summaries. Thus we lose data.

Also, if the process exited before sar ran, there would be no data and some of the generate_system_metrics code would crash expecting to find that process in the sar results.

Fine tuned the process grep to avoid matching processes we don't care about.
Put a catch to not crash if one of the processes we expect doesn't have any data.