(PIE-178) Print metrics as line-delimited JSON
This commit updates the scripts that collect metrics from multiple
hosts, amq_metrics, puma_metrics, and tk_metrics, to write each
dataset to STDOUT with a trailing newline. The resulting output
stream is therefore in line-delimited JSON format, which is easy
to parse.

Prior to this commit, datasets from multiple hosts were mashed together
without separators, which was difficult to parse when used as input
for other scripts.