Fix duplicate declaration of common files
If the `puppet_metrics_collector::system` class is evaluated before the
`puppet_metrics_collector` class, a duplicate declaration error will
occur for output directories and scripts that are common to both classes.
The `system` class has guard logic to prevent this issue, this commit
copies the same logic to the top-level `puppet_metrics_collector` class
and updates the tests to ensure both orderings compile.