(GH-81) Enable client ssl cert for metrics
Prior to this commit, all trapperkeeper metrics were not using the
client ssl certificate. With the changes in PE-28647 to enable
authentication in metrics collection, we now need to leverage the
`auth.conf` rules which require client ssl authentication. This commit
defaults to using the client ssl for all tk metrics.