(GH-81) Handle an empty array when PuppetDB is down
Prior to this commit, the collection would fail with a undefined method
when there was a PuppetDB instance that was offline. This commit adds a
check to ensure that the undefined method error does not happen, while
maintaining the error array. This was reproducable by stopping PuppetDB
and running PuppetDB collection. It would previously fail to do the
collection and throw an exception. After this commnit, the collection
will complete and the error array will be updated with the connection
failure.