(PE-31705) Re-enable remote metric collection
This commit updates the collection of metrics from JVM services,
Puppet Server, PuppetDB, etc., to allow remote collection of
data from the `/metrics/v2/api` if the PE version is 2019.8.5
or newer. Remote collection from this API was disabled due
to the fix for CVE-2020-7943, which caused the API to return
a failure response for requests that did not originate from
localhost.

The restriction may also be removed by setting the following
Hiera parameter, which may be useful for FOSS users or users
on older PE versions:

```
puppet_metrics_collector::pe_metric::remote_metrics_enabled: true
```