Standardize cleanup of temp files
In the case of an error or if the files were otherwise deleted, the
previous commit would try to delete $tmp twice.  This commit adds an
array to hold the temp file names and deletes them on exit, checking the
existence of each one.

Also check if $metrics_tmp is >0 bytes before running xargs.  It could
be empty if there are no files to tidy.