(GH-108) Only manage system collection when sysstat is installed
Prior to this commit, the default was to manage the sysstat package by
default. This commit changes the default to not manage the sysstat
package. It includes logic to only include the system metrics collection
classes when the sysstat package is installed and adds a notify when the
packge is not installed.