(maint) Change default for hosts_with_pe_profile when storeconfigs is false
Previously, this would set the hosts to ['127.0.0.1'] when storeconfigs was false. This would cause flip flops when running puppet infra configure -> puppet agent -t.  This sets the default to the host certname, since this is what the value will end up being on the primary post-install, preventing this flip flop.