Remove tests for "active" metrics services
Prior to this commit, the acceptance tests were testing for either an
active or inactive state for the services.  We don't need to do this for
oneshot services triggered by a timer, since we only care about the
state of the timer.  For PE metrics, we do have a test that explicitly
runs the service to check that the output directory is being populated,
so that case is already accounted for.