(SUP-2725) Fix system collection frequency
Prior to this commit, the postgres and vmware classes didn't correctly
set the 'minute' parameter of the defined type for collecting metrics.
It lacks the "0/" prefix, meaning for a default value of 5 we are
performing the collections on the 5th minute, not repeating every 5
minutes.  This commit adds the prefix.