(SUP-4126) Fix file sync additional metrics
Prior to this commit, the mbeans for the additional file sync storage
metrics started with 'puppetserver:name=puppetlabs.puppet'; however, the
object name of the mbeans include the hostname of the compiling server.
This resulted in the api returning 404s and not collecting the metrics.

This commit replaces the 'puppet' part of the name with a glob, which
should match correctly.