(SUP-4870) Exclude comidi-route-metrics from Console collection
This commit updates the default configuration for Console metrics
collection to exclude the `comidi-route-metrics` dataset.
The console uses heavily parameterized API paths which can result
in this dataset containing several megabytes of information, most
of which is noise. This can consume several hundreds of megabytes
or gigabytes of storage for a day's worth of metrics and can lead
to disk space exhaustion if cleanup is not active.

If needed, this data can be recovered by analyzing the Console
access logs.