(PE-40346) Ensure additional_metrics requests return Array
This commit adds a guard clause to the `retrieve_additional_metrics`
function in `pe_metrics.rb` that ensures the data returned from the
`/metrics` API is of type Array.

Non-conforming data is logged to stderr along with the URL queried
as non-Array responses typically contain error messages that inform
why the request failed.